+++ CHANGELOG.md
@@ -38,6 +38,8 @@
 - [FIX] Wait for disconnect while update
 - [FIX] Fix mfkey32
 - [FIX] Use CurrentActivityHolder for get activity in updater(replace Context cast)
+- [FIX] Back on mfkey32 screen
+- [FIX] Cash with ulong mapping to string in mfkey32
 - [FIX] Move apps from experimental settings to debug
 - [FIX] Fix customize button when set widget state
 - [REFACTOR] Migrate bottom bar to compose navigation